Implement Commit, Duplicate and Export operations
